mwaa-etl-processor/airflow-bucket/requirements.txt
apache-airflow[amazon]==2.8.1
